move google auth to different class
