+++ app/src/main/java/org/fossasia/openevent/general/auth/LoginFragment.kt
@@ -3,14 +3,11 @@ package org.fossasia.openevent.general.auth
 import android.support.v4.app.Fragment
 import android.arch.lifecycle.Observer
 import android.content.Intent
-import android.content.IntentSender
 import android.os.Bundle
-import android.support.v4.app.ActivityCompat
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import com.google.android.gms.auth.api.Auth
 import kotlinx.android.synthetic.main.fragment_login.*
 import kotlinx.android.synthetic.main.fragment_login.view.*
 import org.fossasia.openevent.general.R
@@ -18,9 +15,7 @@ import org.fossasia.openevent.general.utils.Utils
 import org.koin.android.architecture.ext.viewModel
 import com.google.android.gms.common.api.GoogleApiClient
 import com.google.android.gms.auth.api.credentials.Credential
-import com.google.android.gms.common.api.Status
 import org.fossasia.openevent.general.utils.nullToEmpty
-import timber.log.Timber
 
 const val SAVE_DATA: Int = 1
 const val FETCH_DATA: Int = 3
@@ -29,10 +24,9 @@ class LoginFragment : Fragment() {
 
     private val loginActivityViewModel by viewModel<LoginFragmentViewModel>()
     private lateinit var rootView: View
-    var googleApiClient: GoogleApiClient? = null
+    private var googleApiClient: GoogleApiClient? = null
     private lateinit var googleAuthBuilder: GoogleAuthBuilder
     private lateinit var credentialLogin: Credential
-    private var isResolving: Boolean = false
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
@@ -79,14 +73,27 @@ class LoginFragment : Fragment() {
             if (googleAuthBuilder.mode == FETCH_DATA) {
                 googleAuthBuilder.redirectToMain()
             } else if (googleAuthBuilder.mode == SAVE_DATA) {
-                saveCredential(credentialLogin)
+                googleAuthBuilder.saveCredential(credentialLogin)
             }
         })
 
         return rootView
     }
 
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
 
+        if (requestCode == FETCH_DATA) {
+            googleAuthBuilder.mode = FETCH_DATA
+            val credential = data?.getParcelableExtra<Credential>(Credential.EXTRA_KEY)
+            credential?.let {
+                processRetrievedCredential(credential)
+            }
+        } else if (requestCode == SAVE_DATA) {
+            googleAuthBuilder.mode = SAVE_DATA
+            googleAuthBuilder.redirectToMain()
+        }
+    }
 
     fun processRetrievedCredential(credential: Credential) {
         loginActivityViewModel.login(credential.id.nullToEmpty(), credential.password.nullToEmpty())
@@ -98,37 +105,6 @@ class LoginFragment : Fragment() {
         activity?.let { googleApiClient?.stopAutoManage(it) }
         googleApiClient?.disconnect()
         super.onPause()
-    }
-
-    fun resolveResult(status: Status, requestCode: Int) {
-        if (isResolving) {
-            return
-        }
-
-        if (status.hasResolution()) {
-            try {
-                startIntentSenderForResult(status.resolution.intentSender, requestCode, null, 0, 0, 0, null)
-                isResolving = true
-            } catch (e: IntentSender.SendIntentException) {
-                Timber.e(e)
-            }
-        } else {
-            Timber.e("Resolution Failed!")
-        }
-    }
-
-    fun saveCredential(credential: Credential) {
-        isResolving = false
-        Auth.CredentialsApi.save(googleApiClient, credential).setResultCallback { status ->
-
-            if (status.isSuccess) {
-                Timber.d("Credential saved")
-                googleAuthBuilder.redirectToMain()
-            } else {
-                Timber.d("Attempt to save credential failed ${status.statusMessage} ${status.statusCode}")
-                resolveResult(status, SAVE_DATA)
-            }
-        }
     }
 
     override fun onResume() {
